+++ fbw-a32nx/src/systems/fmgc/src/components/fms-messages/FmMessages.ts
@@ -187,4 +187,10 @@ export const FMMessageTypes: Readonly<Record<string, FMMessage>> = {
     color: 'Amber',
     clearable: true,
   },
+  LateralDiscontinuityAhead: {
+    id: 21,
+    text: 'LAT DISCONT AHEAD',
+    color: 'Amber',
+    clearable: true,
+  },
 };