+++ .github/CHANGELOG.md
@@ -27,6 +27,7 @@
 1. [A32NX/MCDU] Fix managed speed units in IAS/MACH on perf climb page - @Lucas-IQ21 (Lucas)
 1. [A32NX/FMS] Implement ability to create, modify, activate and swap a secondary flight plan - @Benjozork (Benjamin Dupont), @BlueberyKing (BlueberryKing)
 1. [A380X] Fix PFD F RELIEF indication and SD status area ISA visibility - @flogross89 (floridude)
+1. [FMS] Add lateral discontinuity ahead fms message - @BravoMike99 (bruno_pt99)
 
 ## 0.14.0